bank/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.6.0               # via kombu
appdirs==1.4.4            # via black, virtualenv
asgiref==3.2.10           # via django
assertpy==1.0             # via -r requirements.in
attrs==19.3.0             # via black, pytest
autoflake==1.3.1          # via -r requirements.in
billiard==3.6.3.0         # via celery
black==19.10b0            # via -r requirements.in
celery==4.4.1             # via -r requirements.in
cffi==1.14.0              # via rabbitmq
click==7.1.2              # via black
distlib==0.3.1            # via virtualenv
django-filter==2.2.0      # via -r requirements.in
django==3.0.4             # via -r requirements.in, django-filter, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
envparse==0.2.0           # via -r requirements.in
factory-boy==2.12.0       # via -r requirements.in
faker==4.1.1              # via factory-boy
filelock==3.0.12          # via tox, virtualenv
freezegun==0.3.15         # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
invoke==1.4.1             # via -r requirements.in
isort==4.3.21             # via -r requirements.in
kombu==4.6.11             # via celery
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest, tox
pathspec==0.8.0           # via black
pluggy==0.13.1            # via pytest, tox
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.9.0                 # via pytest, tox
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via autoflake
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==5.4.3             # via -r requirements.in, pytest-django
python-dateutil==2.8.1    # via faker, freezegun
pytz==2020.1              # via celery, django
rabbitmq==0.2.0           # via -r requirements.in
regex==2020.6.8           # via black
six==1.15.0               # via freezegun, packaging, python-dateutil, tox, virtualenv
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, tox
tox==3.14.5               # via -r requirements.in
typed-ast==1.4.1          # via black
uwsgi==2.0.19.1           # via -r requirements.in
vine==1.3.0               # via amqp, celery
virtualenv==20.0.25       # via tox
wcwidth==0.2.5            # via pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
